// src/types/Student.ts
export interface Student {
  id?: number;
  registerNumber: string;
  studentName: string;
  batchName: string;
  semesterName: string;
  specializationName: string;
  jobPreferredName: string;
  dScore: number;
  iScore: number;
  sScore: number;
  cScore: number;
  assignment: string;
  status: string;
}